Port ElementToScopeBinding to Binding API #37
